requirements.txt

fastapi[standard]
uvicorn
python-dotenv
python-multipart
numpy
huggingface-hub
sentence-transformers
transformers
openai>=1.54.0
anthropic>=0.40.0  # Required for Claude Opus 4.5
vecs
torch
langchain_huggingface
langchain_docling
langchain-community
langchain[openai]
langchain_community
langchain_openai
langchain_anthropic
langchain_text_splitters
langgraph
langsmith
pypdf
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
google-cloud-storage>=2.0.0
sqlalchemy
alembic
pydantic
pydantic-settings
asyncpg
pgvector
greenlet
psycopg2-binary

# --- RAG-specific packages ---
PyMuPDF
redis
Jinja2
docling

# --- Reranking packages ---
cohere>=5.0.0

